Add auto_delete

+++ telefire.py
@@ -1,7 +1,10 @@
+import re
 import sys
 import fire
+import asyncio
 import aiohttp
 import logging
+from datetime import timedelta
 from telethon import utils
 from telethon.sync import TelegramClient, events
 from telethon.tl.types import InputMessagesFilterEmpty
@@ -44,11 +47,47 @@ class Telegram(object):
             utils.get_display_name(user),
             msg.text))
 
+    def _parse_auto_delete_message(self, text):
+        m = re.search(r'^([\d]+[s|m|h|d]) (.*)$', text, re.DOTALL)
+        if m:
+            num = int(m.group(1)[:-1]) if m.group(1)[:-1].isnumeric() else None
+            net = m.group(1)[-1]
+            if num is not None:
+                if net == 'd':
+                    num *= 86400
+                elif net == 'h':
+                    num *= 3600
+                elif net == 'm':
+                    num *= 60
+            return num, m.group(2)
+        return None, None
+
+    async def _auto_delete_async(self, msg, t, text):
+        template = "{}\n==========\nTTL: {}"
+        delta = timedelta(seconds=t)
+        while delta > timedelta(0):
+            await msg.edit(text=template.format(text, str(delta)))
+            if delta.days > 1:
+                await asyncio.sleep(86400)
+                delta -= timedelta(days=1)
+            elif delta.seconds >= 120:
+                await asyncio.sleep(60)
+                delta -= timedelta(seconds=60)
+            elif delta.seconds >= 20:
+                await asyncio.sleep(10)
+                delta -= timedelta(seconds=10)
+            else:
+                await asyncio.sleep(1)
+                delta -= timedelta(seconds=1)
+        await msg.edit(text='[DELETED MESSAGE]')
+        await asyncio.sleep(60)
+        await msg.delete()
+
     async def _get_all_chats(self):
         async for dialog in self._client.iter_dialogs():
             self._logger.info('{:>14}: {}'.format(dialog.id, dialog.title))
 
-    async def _delete_all_async(self, chat=''):
+    async def _delete_all_async(self, chat, query):
         user = await self._client.get_me()
         channel = await self._client.get_entity(chat)
 
@@ -56,8 +95,9 @@ class Telegram(object):
         self._logger.info("Deleting all messages for {} in {}".format(
             utils.get_display_name(user), channel.title))
         async for msg in self._client.iter_messages(channel, from_user=user):
-            self._log_message(msg, channel, user)
-            await msg.delete()
+            if not query or query in msg.text:
+                self._log_message(msg, channel, user)
+                await msg.delete()
 
     async def _list_messages_async(self, chat, user=None):
         channel = await self._client.get_entity(chat)
@@ -117,25 +157,34 @@ class Telegram(object):
             self._client.loop.run_until_complete(
                     self._search_messages_async(peer, query, slow, limit, from_id))
 
-    def list_messages(self, chat='', user=None):
-        if not chat:
-            self._logger.error("Chat cannot be enpty!")
-            return
+    def list_messages(self, chat, user=None):
         with self._client:
             self._client.loop.run_until_complete(
                     self._list_messages_async(chat, user))
 
-    def delete_all(self, chat=''):
-        if not chat:
-            self._logger.error("Chat cannot be enpty!")
-            return
+    def delete_all(self, chat, query=''):
         with self._client:
             self._client.loop.run_until_complete(
-                    self._delete_all_async(chat))
+                    self._delete_all_async(chat, query))
 
     def get_all_chats(self):
         with self._client:
             self._client.loop.run_until_complete(self._get_all_chats())
+
+    def auto_delete(self):
+        @self._client.on(events.NewMessage(pattern=r'^[\d]+[s|m|h|d] '))
+        async def _inner(event):
+            msg = event.message
+            if not msg.out:
+                return
+            channel = await event.get_chat()
+            user = await event.get_sender()
+            self._log_message(msg, channel, user)
+            t, text = self._parse_auto_delete_message(msg.text)
+            await self._auto_delete_async(msg, t, text)
+
+        self._client.start()
+        self._client.run_until_disconnected()
 
 
 if __name__ == '__main__':
